Split core parts of Grep into GrepCore; Use templates to decouple from clp specific implemantions.
