+++ components/core/src/clp/Grep.hpp
@@ -1,10 +1,7 @@
 #ifndef CLP_GREP_HPP
 #define CLP_GREP_HPP
 
-#include <optional>
 #include <string>
-
-#include <log_surgeon/Lexer.hpp>
 
 #include "Defs.h"
 #include "Query.hpp"
@@ -30,60 +27,6 @@ public:
     );
 
     // Methods
-    /**
-     * Processes a raw user query into a Query
-     * @param archive
-     * @param search_string
-     * @param search_begin_ts
-     * @param search_end_ts
-     * @param ignore_case
-     * @param lexer DFA for determining if input is in the schema
-     * @param use_heuristic
-     * @return Query if it may match a message, std::nullopt otherwise
-     */
-    static std::optional<Query> process_raw_query(
-            streaming_archive::reader::Archive const& archive,
-            std::string const& search_string,
-            epochtime_t search_begin_ts,
-            epochtime_t search_end_ts,
-            bool ignore_case,
-            log_surgeon::lexers::ByteLexer& lexer,
-            bool use_heuristic
-    );
-
-    /**
-     * Returns bounds of next potential variable (either a definite variable or a token with
-     * wildcards)
-     * @param value String containing token
-     * @param begin_pos Begin position of last token, changes to begin position of next token
-     * @param end_pos End position of last token, changes to end position of next token
-     * @param is_var Whether the token is definitely a variable
-     * @return true if another potential variable was found, false otherwise
-     */
-    static bool get_bounds_of_next_potential_var(
-            std::string const& value,
-            size_t& begin_pos,
-            size_t& end_pos,
-            bool& is_var
-    );
-
-    /**
-     * Returns bounds of next potential variable (either a definite variable or a token with
-     * wildcards)
-     * @param value String containing token
-     * @param begin_pos Begin position of last token, changes to begin position of next token
-     * @param end_pos End position of last token, changes to end position of next token
-     * @param is_var Whether the token is definitely a variable
-     * @param lexer DFA for determining if input is in the schema
-     * @return true if another potential variable was found, false otherwise
-     */
-    static bool get_bounds_of_next_potential_var(
-            std::string const& value,
-            size_t& begin_pos,
-            size_t& end_pos,
-            bool& is_var,
-            log_surgeon::lexers::ByteLexer& lexer
-    );
     /**
      * Marks which sub-queries in each query are relevant to the given file
      * @param compressed_file
