+++ components/core/src/clp/StringReader.cpp
@@ -7,8 +7,6 @@
 #include <cassert>
 #include <cerrno>
 
-#include <boost/filesystem.hpp>
-
 using std::string;
 
 namespace clp {
